feat: folder structure moved to new component
